Fixed scale uniformly strategy=auto such that it does not waste so much space
it will now apply a similar scheme as earlier for the "optimal
reference" for unit vector ratio.

The idea is to select the strategy
(a) such that it is functionally correct (does not clip away content)
and
(b) such that it does not waste white space (i.e. enlarge limits too
much)