Made a huge step towards backw. compatibility of the new scaling implementation
Now, the scales to individual axes are actually INVERSE scales. This is
necessary because in earlier versions (\pgfplotsset{compat=1.3} to be
more precise), I had \pgfmathdivide{is-width}{should-be-width}. Working
with the actual scales as done in the axis_equal_fix relied on

\pgfmathmultiply{is-width}{inverse of should-be-width}.

This, however, resulted in figure differences of about 2-3pt which seems
to be inacceptable.

Now, I restored \pgfmathdivide where possible.

Note that this did not affect \pgfplotsset{compat=1.5} or newer (because
that relied on float arithmetics anyway).